fix: id type을 int -> Long으로 변환

+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/entity/Bookmark.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 @RequiredArgsConstructor
 public class Bookmark extends BaseEntity {
     @NonNull
-    private int studyId;
+    private Long studyId;
     @NonNull
-    private int userId;
+    private Long userId;
 }
